+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@storefrontui/vue",
-  "version": "0.0.1-aplha.10",
+  "version": "0.0.1-aplha.11",
   "scripts": {
     "serve": "vue-cli-service serve",
     "build": "vue-cli-service build --target lib --name storefront-ui ./src/index.js",
@@ -20,7 +20,6 @@
   "files": [
     "dist/*",
     "scripts/*",
-    "node_modules/*",
     "public/assets"
   ],
   "devDependencies": {
@@ -52,6 +51,9 @@
     "typescript": "^3.4.1",
     "vue-cli-plugin-storybook": "^0.6.0",
     "vue-template-compiler": "^2.6.8"
+  },
+  "peerDependencies": {
+    "vue": "^2.6.x"
   },
   "main": "./dist/storefront-ui.common.js",
   "gitHooks": {
