+++ src/components/organisms/SfAccordion/SfAccordion.html
@@ -1,0 +1,3 @@
+<div>
+  <slot></slot>
+</div>
