+++ src/components/organisms/SfAccordion/SfAccordion.js
@@ -1,0 +1,8 @@
+import Vue from "vue";
+import SfAccordionItem from "@/components/organisms/SfAccordion/_internal/SfAccordionItem.vue";
+
+Vue.component("SfAccordionItem", SfAccordionItem);
+
+export default {
+  name: "SfAccordion"
+};
