+++ src/examples/pages/category/Category.vue
@@ -19,83 +19,85 @@
 
     <div class="main">
       <div class="sidebar desktop-only">
-        <SfAccordion :open="true">
-          <h3 slot="label">Clothing</h3>
-          <SfList slot="content" class="sidebar__category-list">
-            <SfListItem>
-              <SfMenuItem label="All" count="280" />
-            </SfListItem>
-            <SfListItem>
-              <SfMenuItem label="Skirts" count="23" />
-            </SfListItem>
-            <SfListItem>
-              <SfMenuItem label="Sweaters" count="54" />
-            </SfListItem>
-            <SfListItem>
-              <SfMenuItem label="Dresses" count="34" />
-            </SfListItem>
-            <SfListItem>
-              <SfMenuItem label="T-shirts" count="56" />
-            </SfListItem>
-            <SfListItem>
-              <SfMenuItem label="Pants" count="7" />
-            </SfListItem>
-            <SfListItem>
-              <SfMenuItem label="Underwear" count="12" />
-            </SfListItem>
-          </SfList>
-        </SfAccordion>
         <SfAccordion>
-          <h3 slot="label">Accesorries</h3>
-          <SfList slot="content" class="sidebar__category-list">
-            <SfListItem>
-              <SfMenuItem label="All" count="280" />
-            </SfListItem>
-            <SfListItem>
-              <SfMenuItem label="Skirts" count="23" />
-            </SfListItem>
-            <SfListItem>
-              <SfMenuItem label="Sweaters" count="54" />
-            </SfListItem>
-            <SfListItem>
-              <SfMenuItem label="Dresses" count="34" />
-            </SfListItem>
-            <SfListItem>
-              <SfMenuItem label="T-shirts" count="56" />
-            </SfListItem>
-            <SfListItem>
-              <SfMenuItem label="Pants" count="7" />
-            </SfListItem>
-            <SfListItem>
-              <SfMenuItem label="Underwear" count="12" />
-            </SfListItem>
-          </SfList>
-        </SfAccordion>
-        <SfAccordion>
-          <h3 slot="label">Shoes</h3>
-          <SfList slot="content" class="sidebar__category-list">
-            <SfListItem>
-              <SfMenuItem label="All" count="280" />
-            </SfListItem>
-            <SfListItem>
-              <SfMenuItem label="Skirts" count="23" />
-            </SfListItem>
-            <SfListItem>
-              <SfMenuItem label="Sweaters" count="54" />
-            </SfListItem>
-            <SfListItem>
-              <SfMenuItem label="Dresses" count="34" />
-            </SfListItem>
-            <SfListItem>
-              <SfMenuItem label="T-shirts" count="56" />
-            </SfListItem>
-            <SfListItem>
-              <SfMenuItem label="Pants" count="7" />
-            </SfListItem>
-            <SfListItem>
-              <SfMenuItem label="Underwear" count="12" />
-            </SfListItem>
-          </SfList>
+          <SfAccordionItem :open="true">
+            <h3 slot="label">Clothing</h3>
+            <SfList slot="content" class="sidebar__category-list">
+              <SfListItem>
+                <SfMenuItem label="All" count="280" />
+              </SfListItem>
+              <SfListItem>
+                <SfMenuItem label="Skirts" count="23" />
+              </SfListItem>
+              <SfListItem>
+                <SfMenuItem label="Sweaters" count="54" />
+              </SfListItem>
+              <SfListItem>
+                <SfMenuItem label="Dresses" count="34" />
+              </SfListItem>
+              <SfListItem>
+                <SfMenuItem label="T-shirts" count="56" />
+              </SfListItem>
+              <SfListItem>
+                <SfMenuItem label="Pants" count="7" />
+              </SfListItem>
+              <SfListItem>
+                <SfMenuItem label="Underwear" count="12" />
+              </SfListItem>
+            </SfList>
+          </SfAccordionItem>
+          <SfAccordionItem>
+            <h3 slot="label">Accesorries</h3>
+            <SfList slot="content" class="sidebar__category-list">
+              <SfListItem>
+                <SfMenuItem label="All" count="280" />
+              </SfListItem>
+              <SfListItem>
+                <SfMenuItem label="Skirts" count="23" />
+              </SfListItem>
+              <SfListItem>
+                <SfMenuItem label="Sweaters" count="54" />
+              </SfListItem>
+              <SfListItem>
+                <SfMenuItem label="Dresses" count="34" />
+              </SfListItem>
+              <SfListItem>
+                <SfMenuItem label="T-shirts" count="56" />
+              </SfListItem>
+              <SfListItem>
+                <SfMenuItem label="Pants" count="7" />
+              </SfListItem>
+              <SfListItem>
+                <SfMenuItem label="Underwear" count="12" />
+              </SfListItem>
+            </SfList>
+          </SfAccordionItem>
+          <SfAccordionItem>
+            <h3 slot="label">Shoes</h3>
+            <SfList slot="content" class="sidebar__category-list">
+              <SfListItem>
+                <SfMenuItem label="All" count="280" />
+              </SfListItem>
+              <SfListItem>
+                <SfMenuItem label="Skirts" count="23" />
+              </SfListItem>
+              <SfListItem>
+                <SfMenuItem label="Sweaters" count="54" />
+              </SfListItem>
+              <SfListItem>
+                <SfMenuItem label="Dresses" count="34" />
+              </SfListItem>
+              <SfListItem>
+                <SfMenuItem label="T-shirts" count="56" />
+              </SfListItem>
+              <SfListItem>
+                <SfMenuItem label="Pants" count="7" />
+              </SfListItem>
+              <SfListItem>
+                <SfMenuItem label="Underwear" count="12" />
+              </SfListItem>
+            </SfList>
+          </SfAccordionItem>
         </SfAccordion>
       </div>
       <div class="products">
